requirements.txt
Flask~=3.1.0
flask-sqlalchemy
boto3~=1.35.74
pyjwt~=2.10.1
flask-cors~=5.0.0
mysqlclient
pymysql
botocore~=1.35.74
werkzeug~=3.1.3
cryptography
requests~=2.32.3
SQLAlchemy~=2.0.36
python-dotenv~=1.0.1